+++ nav2_bringup/params/nav2_multirobot_params_1.yaml
@@ -95,7 +95,8 @@ bt_navigator:
     - nav2_spin_cancel_bt_node
     - nav2_back_up_cancel_bt_node
     - nav2_drive_on_heading_cancel_bt_node
-    navigator: ["navigate_to_pose", "navigate_through_poses"]
+    navigator: ["navigate_to_pose",
+     "navigate_through_poses"]
     navigate_to_pose:
       plugin: "nav2_bt_navigator/NavigateToPoseNavigator"
