+++ nav2_msgs/action/NavigateToPose.action
@@ -1,6 +1,8 @@
 #goal definition
 geometry_msgs/PoseStamped pose
 string behavior_tree
+string controller_id
+string planner_id
 ---
 #result definition
 std_msgs/Empty result
